Enable and set up Common UI plugin
Enable the Common UI plugin and create some assets needed to make it
work. Included here are controller data assets that define the input
brushes for the interaction keys ("E"/"F" on keyboard and "Y"/"X" on
any XBox-like gamepads). A new "interact" input action is also defined
and a mapping added to the default mapping context. This can be used
instead of the hard-coded keyboard events.

Also introduced, though they are not yet used, are default UI input
actions for "accept" and "cancel". These will later on be needed to
activate and deactivate widgets using Common UI.