Add interaction widget to shop keeper and dialogue base class
Both the conversation dialogue base class and the shop keeper can be
interacted with. In order to make it clearer to the player what they are
expected to do, add an interaction widget that shows the controls to use
and describes the action that will be performed.

Since both the shop keeper and dialogue actors have functionality in
common and will both need the interaction widget, a new interactable
object actor is created and used as the base class for both of them.

This also introduces some basic text and border styles that can be used
as a base line to create consistent UI themes.