Turn party widget into activatable widget
Add some infrastructure to enable working with Common UI's activatable
widgets. To help with this, the HUD adds a layout widget to the viewport
and activatable widgets can be pushed to that layout. Pushing a widget
will automatically activate it, which in turn will set the input mode
and add a mapping context. These menus are automatically cleaned up when
they are deactivated.

Turn the party/pause widget into an activatable widget and push it to
the layout's "game menu" layer when the "pause" input action is
triggered.

Unfortunately there seems to be an issue with Common UI not properly
restoring the previous input mode when deactivating a widget. This has
worked on other projects, so could be an issue in this version of UE. Or
it could be something that I missed. I'll keep investigating as I work
towards converting more widgets.