Use Common UI for start menu
This avoids the need to create a dummy actor for the sole reason to make
it add a widget on screen. Instead we use a separate game mode that sets
a custom HUD class, which derives from the base HUD class used to create
the standard layout widget and then push the start menu to the menu
layer.

Using Common UI for this also takes care of automatically setting the
correct input modes, showing the cursor as necessary and so on.

While at it, set the camera actor to activate for player 0, which makes
sure that it will be used as the view target.