Remake the shop widget using Common UI
The nice bit about doing it this way is that the list view is
dynamically created by simply adding items to it, which moves the
responsibility for creating the list items widgets to the widget
itself.

Since the shop widget is now also an activatable widget, setup and
teardown become significantly simpler because Common UI takes care of
setting the input mode, etc.

As part of this, the BattlePlayerController class is made a child
of the BP_PlayerController class, and the inventory moved to the latter,
which allows the inventory to be accessed more consistently.