Add a tool to recompile Blueprints and their referencers
This is a utility widget that creates a "Scripted Asset Actions" submenu
that, when executed, recompiles the Blueprints selected in the content
drawer as well as all of their referencers. This can be used to quickly
check if any changes to a Blueprint cause other Blueprints to fail
(there is no built-in way to do this, apparently, and play-in-editor
does not reliably compile all Blueprints either).

To use this, run the editor utility by right-clicking and choosing "Run
Editor Utility Object", after which a new sub-menu is added to the right
click context menu of Blueprint assets in the content drawer. Use the
"Scripted Asset Actions" - "Blueprints" - "Recompile all referencers"
option to compile.

Note that this isn't fully integrated into the editor in that any errors
that occur during compilation only show up in the output log and are not
clickable like the errors shown in the compilation results window. But
it should at least give enough pointers to quickly find broken Blueprint
objects.