Add explicit icon textures to abilities and items
Instead of guessing the correct item for abilities and items, add an
explicit reference to each of them. This is slightly more work when
setting up a new ability or item, but it makes the system more robust
against the wrong icon getting picked because of unwanted substring
matching.

While at it, also fix the display of these icons in the combat HUD
widgets (prevent stretching of the icon, etc.).