Rework save system
During testing I noticed that not all information is reliably saved to
the save game. The cause of that seems to be that a number of code sites
load/create and save the game state. This can lead to situations where
not all the relevant information ends up in the save file.

For example, if the game was saved due to the party members getting
updated, but the inventory hadn't been changed yet, a blank inventory
would get saved (and later restored).

Similarly, the player position would usually end up being written as the
world origin and cause the player to get spawned below the landscape.

This reworks ensures that every call site that works with the save game
will use the central save game initialized by the game instance. That
way all data that is modified during the game is always stored in that
central place and the above issues don't happen.

While at it, the player location is now also stored in the save game
data when quitting the game so that players resume from their last
location. In fact, the entire player transform is now stored, so that
the player character also faces the same direction. Not that that's a
big deal, but might as well be as accurate as possible.

Note that this doesn't rework any of the zone save files yet, but they
are less impacted by the kinds of race conditions above, so probably
don't need this kind of rework.