Fix errors in the victory screen
The victory screen currently has a fixed number of party member slots
and this causes errors (on Tick()) when less than three party members
participated in a battle.

Fix this by checking if the unit associated with a slot is valid and
hide it otherwise.

Eventually this should be turned into a list view with a dynamic number
of slots based on the actual number of party members.