Add party member info data assets
Some information is added in both the overworld and battle system party
member blueprints. Attempt to consolidate some of this information into
a separate data asset for each character.

Information such as the battle and overworld classes are also stored so
that eventually we can spawn actors from this shared data.