Use party member info to manage party
Party membership is stored in the save game and the information is
loaded and saved on multiple occasions, and multiple times even.

To reduce this overhead, which can also lead to subtle issues with
consistency, move the party membership information to the player
controller and add some helpers to add, remove, check for membership.
The battle game mode still relies on the save game data, but that's
okay because it doesn't need to modify the membership information.

This gets rid of a bunch of "Get All Actors of Class" nodes used to
determine the party members and generally simplifies the code that
deals with spawning and destroying party members.