Fix the initial player location
When no save game exists, the player is teleported to the world origin
and ends up under the landscape, falling into the endless depths.

Fix this by marking the initial (default) player transform in the save
game as invalid and skip the initial teleport. When saved, the transform
is marked as valid and subsequent loading will respawn the player at the
valid transform.

While at it, remove some code from the game mode that was also trying to
teleport the player. This is already done in the player character, so no
need to repeat it.