Restart player at player start
When resetting the player, teleport to the player start. This is a bit
more friendly than resetting to a hard-coded position and allows the
reset position to be controlled by placing the player start. It's also
more consistent since the reset position is the same as the initial
position of the player.