Add a system screen to the pause menu
The party menu that currently serves as a pause menu does not provide a
way to exit the game, which is somewhat problematic for packaged builds.
Add a switcher and tab list, and complement the pause menu with a system
screen that contains a button quit and resume the game. Quitting will
ask the user whether to quit to the main menu or the desktop.

While at it, also apply the button style used in this pause menu to
other buttons (such as the main menu).

The party selection menu is now also drastically simplified by using a
list view and attaching party member information to each list entry so
that handling of button presses can be moved to the button
implementation, which removes a lot of repetitive code from the party
widget.