Fix interaction widgets popping up erroneously
The overlap check for interactable objects doesn't verify that the
interacting actor is the player character, which results in interaction
widgets becoming visible if any other actor starts to overlap with the
interactable.

Fix this by casting the overlapping actor to the player character first
and ignoring any other events.