Convert dialogue to activatable widget
Entering the dialogue with an NPC keeps the mouse input active and
causes the view to spin around as the player moves the mouse to click
the next button.

Convert the dialogue widget to an activatable widget so that the input
mode is automatically switched to UI. Also make use of existing building
blocks for the user interface to make the dialogue UI look more
consistent with the rest.