Move inventory to actor component
A lot of the inventory functionality can be reused for the shop keepers,
so move it into an actor component that can later be added to the shop
keepers. All paths that use the inventory are updated and the saving of
the inventory is simplified so a function can be called directly on the
component.

The inventory component can add items, remove items and check for the
presence of items.

While at it, create an item to represent gold and integrate it into the
shop widget.

Lastly, add a basic chest interactable object that can be used to test
inventory functionality. It's not currently added to the map to avoid
conflicts.