Reuse inventory component for the shop keeper
By reusing this component, all of the related functionality in widgets
and such can be parameterized and a lot of duplicate code removed. As a
side-effect, this means that the shop inventory can be customized per
instance.

The shop keeper inventory is slightly special in that it doesn't have
any load/save functionality.

While reworking the shop widget, remove any remnants of the old, now
defunct, version of the shop user interface.