Implement item trading in shop keeper UI
Clicking the buy and sell buttons now properly transfers items between
the shop keeper and player inventories and updates the player's amount
of gold.

One potential improvement to this would be to disable buttons for items
that the player cannot click to purchase due to the lack of gold. The
current implementation checks for this and prevents the player from
buying what they cannot afford, but the button can still be clicked,
which may be confusing.