Convert battle HUD to Common UI
Move the top-level HUD widget as well as all of the subwidgets to Common
UI. This simplifies the handling quite a bit and adds gamepad support.

In some cases various smaller bugs were fixed and the magic abilities
and items widgets now work correctly.

This also refactors certain aspects to remove duplicate code. A lot more
could probably be done (magic and item ability slots could potentially
share more code and layout), but this is good enough for now.