Tweak battle HUD show/hide and defend duration
The defend button is slightly special compared to other buttons in that
it requires the information bar to remain visible while the rest of the
battle HUD is disabled (a message is shown for a little bit). In order
to support this, move the show/hide animation for the unit HUD into the
unit HUD widget so that the menu can be controlled separately.

Also slightly tweak the duration that the informational message is show
on screen by moving the delay before the "end turn" call (which results
in a different unit taking over the HUD immediately and taking down the
informational message, giving the player no chance of actually reading
it).