Rework victory screen
Use existing styles to make the victory screen more consistent with the
rest of the UI in the game. Also, turn the party list into a common list
view. This has the advantage of making the list independent of the
number of party members (used to be capped at 3) but comes at the cost
of loosing the animation when the screen is shown. This can probably be
added back in a slightly different way, but that's not addressed here
yet.