Add new interaction system
This new system is attempting to make interactions more universal. The
current system doesn't work because it relies on the interactable actors
to be able to intercept input events from the player controller. This
works fine for regular actors, but anything that derives from the Pawn
class cannot receive input from anything other than it's possessing
controller.

Introduce a new method that relies on interactable objects implementing
a interactable interface. When the player (or more accurately, anything
that contains an interactor component) enters the interactivity sphere
of an interactable object, the object adds itself to the interactor's
list of potential interactables. When the interaction key is pressed,
the closest interactable will be chosen and interacted with.

Note that currently this isn't fully parameterized. The interactor will
only act on the "interact" action event, so if an interactable has a
different input action associated with it, the interaction will still
take place and potentially cause confusion. It should be relatively easy
to fix that up.

An interactable enemy base class is provided that can be used mostly as
a drop-in replacement for the existing enemy base class. However, some
information may be lost during the reparenting, so a better approach may
be to reimplement the new functionality in the old enemy base class.