Explicitly pass through the current unit when creating the magic abilities UI
Instead of relying on the game mode to provide the current unit, pass it
in explicitly when creating the magic abilities UI. This avoids an error
that can occur when the game mode has not selected a current unit yet
when the UI is built.

Note that this really shouldn't happen, so there's probably a subtle bug
somewhere else.