Set the current unit before starting its turn
Some of the battle code relies on the game mode to provide the currently
selected unit. However the current unit isn't set until after the turn
for that unit has been started, which can lead to errors when code wants
to retrieve the unit.

Set the unit before the first code is run to avoid these errors.