Correctly pass the combat usable through to widgets
During the recent rework of the battle HUD, the combat usable wasn't
properly wired through to all widget, effectively giving all abilities
the characteristics of the default attack.

Pass the correct combat usable to all widgets and refactor some code
in the process to make errors less likely to happen.

In particular, this fixes issues with attacks that target all enemies.